Refactor away from trait for model specific implementation
